requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
PyPDF2==3.0.1
beautifulsoup4==4.12.2

# NLP
spacy==3.7.2
sentence-transformers==2.2.2

# Database Clients
neo4j==5.14.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Vector Search
faiss-cpu==1.7.4

# LLM APIs
openai==1.3.7
anthropic==0.7.1

# HTTP
requests==2.31.0
httpx==0.25.2

# Configuration
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Task Scheduling
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Hyperparameter Tuning
optuna==3.4.0

# Retry Logic
tenacity==8.2.3

# Progress Bars
tqdm==4.66.1

# Demo UI
streamlit==1.29.0
plotly==5.18.0
streamlit-agraph==0.0.45

# Deployment
gunicorn==21.2.0
